+++ crates/biome_js_analyze/tests/specs/nursery/noUndeclaredDependencies/valid.d.ts
@@ -1,0 +1,5 @@
+declare module "jest";
+declare module "*.scss" {
+	const content: Record<string, string>;
+	export default content;
+}